Moved MapPauliAxis to Canon
