+++ library/std/qsharp.json
@@ -5,7 +5,6 @@
     "src/legacy_api.qs",
     "src/QIR/Intrinsic.qs",
     "src/Std/Arrays.qs",
-    "src/Std/BlochSphere.qs",
     "src/Std/Canon.qs",
     "src/Std/Convert.qs",
     "src/Std/Diagnostics.qs",
